+++ src/Plainion.Prism/Interactivity/InteractionRequest/AsyncWindowRequestFactory.cs
@@ -3,6 +3,7 @@ using System.ComponentModel.Composition;
 
 namespace Plainion.Prism.Interactivity.InteractionRequest
 {
+    [Obsolete("Prism 7 marked PopupWindowAction as obsolete. Use IDialogService instead")]
     [Export( typeof( IAsyncWindowRequestFactory ) )]
     public class AsyncWindowRequestFactory : IAsyncWindowRequestFactory
     {