+++ src/Plainion.Prism/Interactivity/InteractionRequest/Dialog.cs
@@ -1,7 +1,9 @@
-﻿using Prism.Interactivity.InteractionRequest;
+﻿using System;
+using Prism.Interactivity.InteractionRequest;
 
 namespace Plainion.Prism.Interactivity.InteractionRequest
 {
+    [Obsolete("Prism 7 marked PopupWindowAction as obsolete. Use IDialogService instead")]
     public class Dialog : Confirmation, IDialog
     {
         public double? Width { get; set; }