+++ src/Plainion.Prism/Interactivity/InteractionRequest/ExitWithoutSaveNotification.cs
@@ -5,6 +5,7 @@ using Prism.Interactivity.InteractionRequest;
 
 namespace Plainion.Prism.Interactivity.InteractionRequest
 {
+    [Obsolete("Prism 7 marked PopupWindowAction as obsolete. Use IDialogService instead")]
     public class ExitWithoutSaveNotification : Notification, IInteractionRequestAware
     {
         public enum ResponseType