+++ src/Plainion.Prism/Interactivity/InteractionRequest/IAsyncWindowRequestFactory.cs
@@ -2,6 +2,7 @@
 
 namespace Plainion.Prism.Interactivity.InteractionRequest
 {
+    [Obsolete("Prism 7 marked PopupWindowAction as obsolete. Use IDialogService instead")]
     public interface IAsyncWindowRequestFactory
     {
         IAsyncWindowRequest CreateForView( Type windowContent );