+++ src/Plainion.Prism/Interactivity/InteractionRequest/IDialog.cs
@@ -1,7 +1,9 @@
-﻿using Prism.Interactivity.InteractionRequest;
+﻿using System;
+using Prism.Interactivity.InteractionRequest;
 
 namespace Plainion.Prism.Interactivity.InteractionRequest
 {
+    [Obsolete("Prism 7 marked PopupWindowAction as obsolete. Use IDialogService instead")]
     interface IDialog : IConfirmation
     {
         double? Width { get; set; }