+++ src/Plainion.Prism/Interactivity/InteractionRequest/SaveFileDialogNotification.cs
@@ -1,6 +1,9 @@
 ﻿
+using System;
+
 namespace Plainion.Prism.Interactivity.InteractionRequest
 {
+    [Obsolete("Prism 7 marked PopupWindowAction as obsolete. Use IDialogService instead")]
     public class SaveFileDialogNotification : FileDialogNotificationBase
     {
         /// <see cref="P:Microsoft.Win32.FileDialog.CreatePrompt"/>