+++ src/Plainion.Prism/Interactivity/InteractionRequest/SelectFolderDialogNotification.cs
@@ -3,6 +3,7 @@ using Prism.Interactivity.InteractionRequest;
 
 namespace Plainion.Prism.Interactivity.InteractionRequest
 {
+    [Obsolete("Prism 7 marked PopupWindowAction as obsolete. Use IDialogService instead")]
     public class SelectFolderDialogNotification : Confirmation
     {
         /// <see cref="P:Plainion.Windows.Controls.SelectFolderDialog.Description"/>