+++ src/Plainion.Prism/Interactivity/PopupCommonDialogAction.cs
@@ -7,6 +7,7 @@ using Microsoft.Win32;
 
 namespace Plainion.Prism.Interactivity
 {
+    [Obsolete("Prism 7 marked PopupWindowAction as obsolete. Use IDialogService instead")]
     public class PopupCommonDialogAction : TriggerAction<FrameworkElement>
     {
         public static readonly DependencyProperty FileDialogTypeProperty =