+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 - Removed unnecessary dependency to Prism.MEF
 - Convenience events removed (ApplicationReadyEvent, ApplicationShutdownEvent)
+- Marked all InteractRequest related classes as obsolete as Prism 7 marks those as obsolete as well
 
 ## 4.1.0 - 2017-11-15